pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = "Python microservice template based on FastAPI framework"
authors = ["Sohail Chamadia <sohail.chamadia@everestek.com>"]

[tool.poetry.dependencies]
python = "^3.9.7"
fastapi = "^0.70.0"
alembic = "^1.7.4"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
requests = "^2.26.0"
fastapi-utils = "^0.2.1"
PyMySQL = "^1.0.2"
fastapi-login = "^1.7.3"
python-decouple = "^3.6"
bcrypt = "^3.2.0"
python-dotenv = "^0.19.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.9.0"}
PyJWT = "^2.3.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.10b0"
isort = "^5.9.3"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^6.2.5"
sqlalchemy-stubs = "^0.4"
commitizen = "^2.20.0"
pre-commit = "^2.15.0"
nox = "^2021.10.1"

[tool.commitizen]
name = "cz_conventional_commits"
version_files = [
    "pyproject.toml:version"
]
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]
bump_message = "release $current_version → $new_version [skip-ci]"
update_changelog_on_bump = true
annotated_tag = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
